requirements-dev.txt
alabaster>=0.7.12
appdirs>=1.4.4
attrs>=19.3.0
Babel>=2.9.1
black>=19.10b0
certifi>=2020.12.5
chardet>=4.0.0
click>=7.1.2
colorama>=0.4.4
docutils>=0.16
fake-bpy-module-2.82>=20200514
flake8>=3.8.1
flake8-black>=0.1.2
flake8-bugbear>=20.1.4
idna>=2.10
imagesize>=1.2.0
importlib-metadata>=1.6.0
Jinja2>=2.11.3
MarkupSafe>=1.1.1
mccabe>=0.6.1
packaging>=20.9
pathspec>=0.8.0
pycodestyle>=2.6.0
pyflakes>=2.2.0
Pygments>=2.8.1
pyparsing>=2.4.7
PySide2>=5.15.2
pytz>=2021.1
regex>=2021.11.2
requests>=2.25.1
snowballstemmer>=2.1.0
Sphinx>=3.5.4
sphinx-rtd-theme>=0.5.2
sphinxcontrib-applehelp>=1.0.2
sphinxcontrib-devhelp>=1.0.2
sphinxcontrib-htmlhelp>=1.0.3
sphinxcontrib-jsmath>=1.0.1
sphinxcontrib-qthelp>=1.0.3
sphinxcontrib-serializinghtml>=1.1.4
toml>=0.10.1
typed-ast>=1.4.1
typing>=3.7.4.1
urllib3>=1.26.6
zipp>=3.1.0
